requirements-minimal.txt
# DurgasAI - Minimal Requirements for Basic Functionality
# Core Streamlit
streamlit>=1.28.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.38
langchain-huggingface>=0.0.3

# HuggingFace ecosystem
transformers>=4.35.0
huggingface-hub>=0.19.0
tokenizers>=0.15.0

# PyTorch (CPU version - lighter)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HTTP requests and API handling
requests>=2.28.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Progress bars
tqdm>=4.65.0

# Date and time
python-dateutil>=2.8.0
